🐛amp-carousel: move mutate from onLayoutMeasure to layoutCallback. (#20336)
The onLayoutMeasure callback could fire during scrolling, causing the scroll position to incorrectly be set. Move the setting of the scroll position into layoutCallback. Since BaseCarousel implements isRelayoutNeeded, returning true, the layoutCallback will be called whenever the size of the carousel itself changes (whether on orientation or otherwise).

Remove forced style calculation by reading the scrollLeft immediately after setting it. Instead, used the value we just requested to be set. The only situation where this would not work is if the requested scrollLeft was larger than the available scrollWidth. To cover this case, added a check to getScrollLeftForIndex to return 0 when there is only 1 slide. As a result, all other cases should have a scrollWidth of at least slideWidth.