✨Create new extension - AMP-smartlinks (#20494)
Resolves https://github.com/ampproject/amphtml/issues/19470

## What it does
Create an extension that lets Narrativ offer our publishers our Linkmate service in AMP pages. The extension scans the page for any outbound links the publisher wants targeted and sends them to our API for monetization. Upon API response we use the `link-rewriter` service to map that response to the links in the initial payload. When a user clicks one of the monetized links we swap the href with the monetized link. More about Linkmate [here](http://docs.narrativ.com/en/stable/linkmate.html#linkmate).