Make inabox-viewport measure the top level window directly if same domain (#20459)
* Make inabox-viewport measure the top level window directly if same domain

Instead of sending a post message, if the inabox iframe can directly measure the host window then have it get its position directly from there.

* fix types

* Fix presubmit

* Test disconnect function

* Fix lint in new test

* Try enabling friendly, no host script inabox test on Safari

* Trivial fix to rerun Travis

* Try the no host script test with the example AMP ad

Checking if the serve mode of the example AMP test is changed somehow

* don't whitelist the iframe

* More work

* Fix lint

* Fix lint in test

* fix oops

* Throttle the listeners

* Use the existing inabox host observer instead of rewriting a new one

This makes the code cleaner and changes in the future easier

* fix type issue

* Revert the changes to test-amp-inabox.js

I believe that the while the integration tests do test the inabox host script, they're are not actually testing the local AMP inabox run time. This might have to be fixed in a separate PR.

* Gate the behavior behind experiment, and move canInspectWindow to helper file

* Integrate the changes in #20599, and make sure the position observer is not null.

* Inline the function

* Fix type

* Disconnect observer properly, and add a TODO to investigate a quirk in the listener.

Specifically, why is Resources needed.