amp-geo: Fix inconsistent geoState parent (#21293)
* amp-geo: Fix inconsistent geoState parent

PR #21282 introduced a new issue where the #ampGeo <amp-state> would be
appended to the ampdoc body but searched for in the win.doc body. Look
for #ampGeo in ampdoc.

* amp-geo: Apply clearPreRender to AMP body

PR #21282 changed the location of amp-geo related classes to the AMP
document body. Propagate the change to the clearPreRender_method as
well.

* amp-geo: Consistently use ampdoc in addToBody_

It is not necessary to retain both doc and ampdoc in the addToBody_
method. Update to use only ampdoc.